feat(linux): Add linux build targets. Add linux build targets similar to those of iob-soc-sut repository. DMA transfers are still not working well. They do work on some runs but not in others. It seems to work when the DMA (dynamically allocaed) shared memory address is less than 0x81000000. Maybe some hardware issue with values above 24 bits?
