+++ Makefile
@@ -66,15 +66,19 @@ test-all:
 
 test-linux-fpga-connect: build_dir_name
 	-rm $(BUILD_DIR)/hardware/fpga/test.log
-	-ln -s minicom_test1.txt $(BUILD_DIR)/hardware/fpga/minicom_linux_script.txt
+	-ln -fs minicom_test1.txt $(BUILD_DIR)/hardware/fpga/minicom_linux_script.txt
 	make fpga-connect RUN_LINUX=1 BOOT_FLOW=$(BOOT_FLOW)
 
 dma-linux-fpga-connect: build_dir_name
 	-rm $(BUILD_DIR)/hardware/fpga/test.log
-	-ln -s minicom_dma_test.txt $(BUILD_DIR)/hardware/fpga/minicom_linux_script.txt
+	-ln -fs minicom_dma_test.txt $(BUILD_DIR)/hardware/fpga/minicom_linux_script.txt
 	make fpga-connect RUN_LINUX=1 BOOT_FLOW=$(BOOT_FLOW)
 
 .PHONY: setup sim-run sim-test fpga-test test-all test-linux-fpga-connect dma-linux-fpga-connect
+
+#
+# Linux targets
+#
 
 LINUX_OS_DIR ?= submodules/OS
 # Relative path from OS directory to OpenCryptoLinux (OCL) directory
@@ -92,16 +96,76 @@ build-linux-opensbi:
 
 .PHONY: build-linux-dts build-linux-opensbi
 
+COMBINED_BUILDROOT_DIR=..
+combine-buildroot:
+	rm -rf $(COMBINED_BUILDROOT_DIR)/buildroot
+	cp -r $(LINUX_OS_DIR)/software/buildroot $(COMBINED_BUILDROOT_DIR)/
+	cp -r ./software/buildroot $(COMBINED_BUILDROOT_DIR)/
+
+build-linux-buildroot: combine-buildroot build-linux-drivers build-linux-dma-demo
+	make -C $(LINUX_OS_DIR) build-buildroot OS_SUBMODULES_DIR=$(REL_OS2OCL)/.. OS_SOFTWARE_DIR=../`realpath $(COMBINED_BUILDROOT_DIR) --relative-to=..` OS_BUILD_DIR=$(REL_OS2OCL)/software/src
+
+.PHONY: combine-buildroot build-linux-buildroot
+
+# System peripherals
 MODULE_NAMES += iob_dma
 MODULE_NAMES += iob_axistream_in
 MODULE_NAMES += iob_axistream_out
+MODULE_NAMES += iob_timer
 
 build-driver-headers:
 	@$(foreach module,$(MODULE_NAMES), \
-		./$(LINUX_OS_DIR)/scripts/drivers.py $(module) -o software/tests/dma/src/; \
+		./$(LINUX_OS_DIR)/scripts/drivers.py $(module) -o `realpath $(COMBINED_BUILDROOT_DIR)`/buildroot/board/IObundle/iob-soc/rootfs-overlay/root/dma_demo/; \
 	)
 
 .PHONY: build-driver-headers
+
+build-linux-drivers:
+	@$(foreach module,$(MODULE_NAMES), \
+		printf "\n\n\nMaking $(module)\n\n\n"; \
+		$(eval MODULE_LINUX_DIR=$(shell realpath $(shell find . -type f -name '$(module).py' -printf '%h' -quit))/software/linux) \
+		nix-shell $(LINUX_OS_DIR)/default.nix --run 'make -C $(LINUX_OS_DIR) build-linux-drivers \
+			MODULE_DRIVER_DIR=$(MODULE_LINUX_DIR)/drivers \
+			OS_SUBMODULES_DIR=$(REL_OS2OCL)/.. \
+			CALLING_DIR=`realpath $(CURDIR)` \
+			MODULE_NAME=$(module) \
+			ROOTFS_OVERLAY_DIR=`realpath $(COMBINED_BUILDROOT_DIR)`/buildroot/board/IObundle/iob-soc/rootfs-overlay/ \
+			PYTHON_DIR=`realpath $(LIB_DIR)`/scripts'; \
+	)
+
+
+clean-linux-drivers:
+	@$(foreach module,$(MODULE_NAMES), \
+		$(eval MODULE_LINUX_DIR=$(shell realpath $(shell find . -type f -name '$(module).py' -printf '%h' -quit))/software/linux) \
+		nix-shell $(LINUX_OS_DIR)/default.nix --run 'make -C $(LINUX_OS_DIR) clean-linux-drivers \
+			MODULE_DRIVER_DIR=$(MODULE_LINUX_DIR)/drivers \
+			OS_SUBMODULES_DIR=$(REL_OS2OCL)/.. \
+			CALLING_DIR=`realpath $(CURDIR)` \
+			MODULE_NAME=$(module) \
+			ROOTFS_OVERLAY_DIR=`realpath $(COMBINED_BUILDROOT_DIR)`/buildroot/board/IObundle/iob-soc/rootfs-overlay/ \
+			PYTHON_DIR=`realpath $(LIB_DIR)`/scripts'; \
+	)
+
+.PHONY: build-linux-drivers clean-linux-drivers
+
+build-linux-kernel:
+	-rm ../linux-5.15.98/arch/riscv/boot/Image
+	nix-shell $(LINUX_OS_DIR)/default.nix --run 'make -C $(LINUX_OS_DIR) build-linux-kernel OS_SUBMODULES_DIR=$(REL_OS2OCL)/.. OS_SOFTWARE_DIR=../`realpath . --relative-to=..`/software OS_BUILD_DIR=$(REL_OS2OCL)/software/src'
+
+.PHONY: build-linux-kernel
+
+# Main target to build all linux related files
+build-linux-files:
+	make build-linux-dts
+	make build-linux-opensbi
+	make build-linux-kernel
+	make build-linux-buildroot
+
+.PHONY: build-linux-files
+
+#
+# Linux software targets
+#
 
 INCLUDE = -I.
 SRC = *.c
@@ -110,10 +174,10 @@ FLAGS += -Werror
 #FLAGS += -static
 FLAGS += -march=rv32imac
 FLAGS += -mabi=ilp32
-BIN = ../dma_demo
+BIN = dma_demo
 CC = riscv64-unknown-linux-gnu-gcc
 build-linux-dma-demo: build-driver-headers
-	nix-shell $(LINUX_OS_DIR)/default.nix --run 'cd software/tests/dma/src && \
+	nix-shell $(LINUX_OS_DIR)/default.nix --run 'cd $(COMBINED_BUILDROOT_DIR)/buildroot/board/IObundle/iob-soc/rootfs-overlay/root/dma_demo/ && \
 	$(CC) $(FLAGS) $(INCLUDE) -o $(BIN) $(SRC)'
 
 .PHONY: build-linux-dma-demo
