+++ software/linux_config
@@ -1,0 +1,84 @@
+CONFIG_SECTION_MISMATCH_WARN_ONLY=y
+
+# Architecture
+CONFIG_ARCH_DEFCONFIG="arch/riscv/configs/defconfig"
+CONFIG_NONPORTABLE=y
+CONFIG_ARCH_RV32I=y
+CONFIG_RISCV_ISA_M=y
+CONFIG_RISCV_ISA_A=y
+CONFIG_RISCV_ISA_C=y
+CONFIG_SIFIVE_PLIC=y
+CONFIG_FPU=n
+CONFIG_SMP=y
+CONFIG_STRICT_KERNEL_RWX=n
+CONFIG_EFI=n
+CONFIG_HVC_RISCV_SBI=y
+CONFIG_HOTPLUG_CPU=y
+
+CONFIG_BLK_DEV_INITRD=y
+CONFIG_INITRAMFS_SOURCE=""
+CONFIG_RD_GZIP=y
+CONFIG_RD_BZIP2=y
+CONFIG_RD_LZMA=y
+CONFIG_RD_XZ=y
+CONFIG_RD_LZO=y
+CONFIG_RD_LZ4=y
+
+# FPGA / SoC
+CONFIG_FPGA=y
+
+# Time
+CONFIG_PRINTK_TIME=y
+
+# Clocking
+CONFIG_COMMON_CLK=y
+CONFIG_COMMON_CLK_LITEX=y
+
+# Interrupts
+CONFIG_IRQCHIP=y
+CONFIG_OF_IRQ=y
+CONFIG_HANDLE_DOMAIN_IRQ=y
+
+# Modules
+CONFIG_MODULES=y
+CONFIG_MODULE_UNLOAD=y
+
+# Ethernet
+CONFIG_NET=y
+CONFIG_PACKET=y
+CONFIG_PACKET_DIAG=y
+CONFIG_INET=y
+CONFIG_NETDEVICES=n
+CONFIG_ETHOC=n
+
+## Serial
+CONFIG_SERIAL_EARLYCON_RISCV_SBI=y
+# Enable UART16550 module and probe them from device tree
+CONFIG_SERIAL_8250=y
+#CONFIG_SERIAL_8250_CONSOLE=y
+CONFIG_SERIAL_OF_PLATFORM=y
+
+CONFIG_EXT2_FS=y
+CONFIG_EXT3_FS=y
+CONFIG_EXT4_FS=y
+
+## .config in kernel
+CONFIG_IKCONFIG=y
+CONFIG_IKCONFIG_PROC=y
+
+## Filesystem
+CONFIG_FAT_FS=y
+CONFIG_MSDOS_FS=y
+CONFIG_MSDOS_PARTITION=y
+CONFIG_VFAT_FS=y
+CONFIG_FAT_DEFAULT_CODEPAGE=437
+CONFIG_FAT_DEFAULT_IOCHARSET="iso8859-1"
+CONFIG_NCPFS_SMALLDOS=y
+CONFIG_NLS=y
+CONFIG_NLS_CODEPAGE_437=y
+CONFIG_NLS_ASCII=y
+CONFIG_NLS_ISO8859_1=y
+CONFIG_TMPFS=y
+CONFIG_DEVTMPFS=y
+CONFIG_DEVTMPFS_MOUNT=y
+CONFIG_UEVENT_HELPER=y
