--- software/tests/dma/src/iob_axistream_in.h
@@ -1,54 +1,0 @@
-/* This file was automatically generated by:
- * `write_linux_swheader` method of `mkregs.py`
- */
-
-#ifndef H_IOB_AXISTREAM_IN_H
-#define H_IOB_AXISTREAM_IN_H
-
-#define IOB_AXISTREAM_IN_DRIVER_NAME "iob_axistream_in"
-#define IOB_AXISTREAM_IN_DRIVER_CLASS "iob_axistream_in"
-#define IOB_AXISTREAM_IN_DEVICE_FILE "/dev/iob_axistream_in"
-#define IOB_AXISTREAM_IN_DEVICE_CLASS "/sys/class/" IOB_AXISTREAM_IN_DRIVER_CLASS "/" IOB_AXISTREAM_IN_DRIVER_NAME
-
-#define IOB_AXISTREAM_IN_SYSFILE_SOFT_RESET IOB_AXISTREAM_IN_DEVICE_CLASS "/soft_reset"
-#define IOB_AXISTREAM_IN_SYSFILE_ENABLE IOB_AXISTREAM_IN_DEVICE_CLASS "/enable"
-#define IOB_AXISTREAM_IN_SYSFILE_DATA IOB_AXISTREAM_IN_DEVICE_CLASS "/data"
-#define IOB_AXISTREAM_IN_SYSFILE_MODE IOB_AXISTREAM_IN_DEVICE_CLASS "/mode"
-#define IOB_AXISTREAM_IN_SYSFILE_NWORDS IOB_AXISTREAM_IN_DEVICE_CLASS "/nwords"
-#define IOB_AXISTREAM_IN_SYSFILE_TLAST_DETECTED IOB_AXISTREAM_IN_DEVICE_CLASS "/tlast_detected"
-#define IOB_AXISTREAM_IN_SYSFILE_FIFO_FULL IOB_AXISTREAM_IN_DEVICE_CLASS "/fifo_full"
-#define IOB_AXISTREAM_IN_SYSFILE_FIFO_EMPTY IOB_AXISTREAM_IN_DEVICE_CLASS "/fifo_empty"
-#define IOB_AXISTREAM_IN_SYSFILE_FIFO_THRESHOLD IOB_AXISTREAM_IN_DEVICE_CLASS "/fifo_threshold"
-#define IOB_AXISTREAM_IN_SYSFILE_FIFO_LEVEL IOB_AXISTREAM_IN_DEVICE_CLASS "/fifo_level"
-#define IOB_AXISTREAM_IN_SYSFILE_VERSION IOB_AXISTREAM_IN_DEVICE_CLASS "/version"
-
-//used address space width
-#define IOB_AXISTREAM_IN_SWREG_ADDR_W 5
-
-//Addresses
-#define IOB_AXISTREAM_IN_SOFT_RESET_ADDR 0
-#define IOB_AXISTREAM_IN_ENABLE_ADDR 1
-#define IOB_AXISTREAM_IN_DATA_ADDR 4
-#define IOB_AXISTREAM_IN_MODE_ADDR 8
-#define IOB_AXISTREAM_IN_NWORDS_ADDR 12
-#define IOB_AXISTREAM_IN_TLAST_DETECTED_ADDR 16
-#define IOB_AXISTREAM_IN_FIFO_FULL_ADDR 17
-#define IOB_AXISTREAM_IN_FIFO_EMPTY_ADDR 18
-#define IOB_AXISTREAM_IN_FIFO_THRESHOLD_ADDR 20
-#define IOB_AXISTREAM_IN_FIFO_LEVEL_ADDR 24
-#define IOB_AXISTREAM_IN_VERSION_ADDR 28
-
-//Data widths (bit)
-#define IOB_AXISTREAM_IN_SOFT_RESET_W 8
-#define IOB_AXISTREAM_IN_ENABLE_W 8
-#define IOB_AXISTREAM_IN_DATA_W 32
-#define IOB_AXISTREAM_IN_MODE_W 8
-#define IOB_AXISTREAM_IN_NWORDS_W 32
-#define IOB_AXISTREAM_IN_TLAST_DETECTED_W 8
-#define IOB_AXISTREAM_IN_FIFO_FULL_W 8
-#define IOB_AXISTREAM_IN_FIFO_EMPTY_W 8
-#define IOB_AXISTREAM_IN_FIFO_THRESHOLD_W 32
-#define IOB_AXISTREAM_IN_FIFO_LEVEL_W 32
-#define IOB_AXISTREAM_IN_VERSION_W 16
-
-#endif // H_IOB_AXISTREAM_IN_H
