--- software/tests/dma/src/iob_dma.h
@@ -1,42 +1,0 @@
-/* This file was automatically generated by:
- * `write_linux_swheader` method of `mkregs.py`
- */
-
-#ifndef H_IOB_DMA_H
-#define H_IOB_DMA_H
-
-#define IOB_DMA_DRIVER_NAME "iob_dma"
-#define IOB_DMA_DRIVER_CLASS "iob_dma"
-#define IOB_DMA_DEVICE_FILE "/dev/iob_dma"
-#define IOB_DMA_DEVICE_CLASS "/sys/class/" IOB_DMA_DRIVER_CLASS "/" IOB_DMA_DRIVER_NAME
-
-#define IOB_DMA_SYSFILE_BASE_ADDR IOB_DMA_DEVICE_CLASS "/base_addr"
-#define IOB_DMA_SYSFILE_TRANSFER_SIZE IOB_DMA_DEVICE_CLASS "/transfer_size"
-#define IOB_DMA_SYSFILE_DIRECTION IOB_DMA_DEVICE_CLASS "/direction"
-#define IOB_DMA_SYSFILE_INTERFACE_NUM IOB_DMA_DEVICE_CLASS "/interface_num"
-#define IOB_DMA_SYSFILE_READY_R IOB_DMA_DEVICE_CLASS "/ready_r"
-#define IOB_DMA_SYSFILE_READY_W IOB_DMA_DEVICE_CLASS "/ready_w"
-#define IOB_DMA_SYSFILE_VERSION IOB_DMA_DEVICE_CLASS "/version"
-
-//used address space width
-#define IOB_DMA_SWREG_ADDR_W 4
-
-//Addresses
-#define IOB_DMA_BASE_ADDR_ADDR 0
-#define IOB_DMA_TRANSFER_SIZE_ADDR 4
-#define IOB_DMA_DIRECTION_ADDR 8
-#define IOB_DMA_INTERFACE_NUM_ADDR 10
-#define IOB_DMA_READY_R_ADDR 12
-#define IOB_DMA_READY_W_ADDR 13
-#define IOB_DMA_VERSION_ADDR 14
-
-//Data widths (bit)
-#define IOB_DMA_BASE_ADDR_W 32
-#define IOB_DMA_TRANSFER_SIZE_W 32
-#define IOB_DMA_DIRECTION_W 8
-#define IOB_DMA_INTERFACE_NUM_W 16
-#define IOB_DMA_READY_R_W 8
-#define IOB_DMA_READY_W_W 8
-#define IOB_DMA_VERSION_W 16
-
-#endif // H_IOB_DMA_H
